requirements.txt
deep-translator==1.4.3
python-dotenv==0.19.1
requests==2.26.0
selenium==4.19.0
Flask==2.0.2
bs4==0.0.1
googletrans==3.1.0a0

Werkzeug>=2.2,<3.0
gunicorn

